requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-engineio==3.9.3
python-socketio==4.3.1
six==1.12.0
typing-extensions==3.7.4
yarl==1.3.0
redis==2.10.3